requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
plotly>=5.18.0
mplfinance>=0.12.10b0
chronos-forecasting>=2.0.0
torch>=2.0.0
accelerate>=0.25.0
transformers>=4.49.0
safetensors>=0.4.0
